low level
low level

+++ 0x13-more_singly_linked_lists/10-delete_nodeint.c
@@ -1,15 +1,17 @@
 #include "lists.h"
+
 /**
  * delete_nodeint_at_index - deletes a node in a linked list at a certain index
- * @head: pointer to the first node in the list
+ * @head: pointer to the first element in the list
  * @index: index of the node to delete
+ *
  * Return: 1 (Success), or -1 (Fail)
  */
 int delete_nodeint_at_index(listint_t **head, unsigned int index)
 {
 	listint_t *temp = *head;
 	listint_t *current = NULL;
-	size_t i = 0;
+	unsigned int i = 0;
 
 	if (*head == NULL)
 		return (-1);
@@ -25,10 +27,10 @@ int delete_nodeint_at_index(listint_t **head, unsigned int index)
 	{
 		if (!temp || !(temp->next))
 			return (-1);
-
 		temp = temp->next;
 		i++;
 	}
+
 
 	current = temp->next;
 	temp->next = current->next;
